[!12][SPEECHFORMER] Add argument to control CTC compression max output length
# Why is the change needed?

The CTC compression does not provide any guarantee on the resulting compression factor and hence sequence length. Especially in early phases of the training, this can cause OOM issues when training the Speechformer, as the length of the input may not be enough reduced to be processed by Transformer layers.

# What changes does the change introduce?

This change introduces a new argument that allows the user to control the maximum output length after the CTC compression by merging consecutive elements, even though they have different CTC predictions, in case the constraint is violated.

# How was the change tested?

Added UTs.