[!11][SPEECHFORMER] Added variable number of convolutional layers
It is not possible to specify how many convolutional layers we want to use for the K and V compression in the Speechformer architecture.

It adds the possibility to vary the number of convolutional layers in the K and V compression to the Speechformer architecture. The default value is 1, as in the published version.

UTs and manual runs on the cluster on our Speechformer en-de model with an increased BLEU score in presence of 2 convolutional layers (+1.03 wrt to the baseline).