[!19][DEV] Script that automatically merges MR with correct formatting
# Why is the change needed?

Correctly merging a MR with all correct format rules can be difficult for non-experts and is error-prone.
A script that automatically does the job enforcing the correct format both reduces the workload
for merging a MR and enforces the correct behavior.

# What changes does the patch introduce?

Adds a script to merge MRs.

# How was this patch tested?

Manual test merging MR !14.