[!15][NE] Read datasets with tagged strings and add weighted loss
# Why is the change needed?

The goal is to work with named entities, terminology, that can be marked in the text for many reasons, either to produce an output with the tagging of these elements or to process them with dedicated care. In the future, there may be other use cases requiring tagged elements.

# What changes does the patch introduce?

The MR adds a dataset loader that processes tagged strings and returns the strings and the tagged tokens.
This enables models and criteria to access these two disentangled data.

# How was this patch tested?

Added UTs.