[!22][DEV] Add CI
<!--
Please clarify why the changes are needed. For instance,
  1. If you propose a new feature, clarify the use case for the new feature.
  2. If you fix a bug, you can clarify why it is a bug.
-->
Currently we do not have tests executed in CI, but we need to manually run tests before merging.

<!--
Please clarify what changes you are proposing.
The purpose of this section is to outline the changes and how this PR fixes the issue.
If possible, please consider writing useful notes for better and faster reviews in your PR.
-->
The patch add CI, so tests are automatically triggered for merge requests
and on the master branch.

<!--
If tests were added, say they were added here.
Please make sure to add some test cases that check the changes thoroughly
including negative and positive cases if possible.
If it was tested in a way different from regular unit tests,
please clarify how you tested step by step, ideally copy and paste-able,
so that other reviewers can test and check, and descendants can verify in the future.
If tests were not added, please describe
why they were not added and/or why it was difficult to add.
-->
manual tests on this branch