[!13][CTC] Add beam search decoding on CTC module in encoder-decoder architectures
# Why is the change needed?

We currently train our models to optimize a multi-task loss that exploits an additional CTC loss in the encoder layers to improve performance/avoid ASR pertaining. However, we have no access to the CTC generated outputs, which may be interesting both for model analysis and for possible other tasks.

A naive greedy search produces highly sub-optimal results, so we need a beam search over the CTC module to obtain meaningful outputs.

# What changes does the change introduce?

It introduces a new task that can be used to generate the outputs using the CTC module of encoder-decoder models.

# How was the change tested?

Manual runs on the cluster on our Speechformer en-de model (see EMNLP 2021) that resulted in a reasonable 18.52 WER.