[!26][CTC] Log metrics of the underlying criterion of CTC multi loss
# Why is the change needed?

The CTC multi-loss criterion does not correctly log all the metrics of the underlying criterion unless the underlying criterion is a cross-entropy (with or without label smoothing). The recent introduction of the CTC support for triangle models (!23) exposed this issue.

# What changes does the patch introduce?

It adds the logging of all metrics of the underlying criterion and delegates to it the metrics reduction.

# How was this patch tested?

Added UTs.