[!33][CTC] Add fixed compression method
# Why is the change needed?

Some architectures, like the Speechformer, rely on the compression produced by the CTC, which - in early stages - is random and can lead to too long/short outputs causing OOM errors. For this reason, we need a method that can ensure that the length of the sequences after compression is reasonable.

# What changes does the patch introduce?

The patch introduces a `fixed` mechanism that compresses the input by a factor of 4.

# How was this patch tested?

UT