[!37][CONFORMER] Add the possibility of changing SyncBatchNorm with BatchNorm1D
# Why is the change needed?
When Conformer-based architectures are trained on a single GPU, an error is raised caused by `SyncBatchNorm` that, instead, requires more than 1 GPU to be used. This error can be solved by adopting `BatchNorm1D` that, however, cannot be used if at least one mini-batch is of size 1. For this reason, we have 3 scenarios:

 - more than 1 GPU: `SyncBatchNorm` works correctly in any case;
 - 1 GPU with enough RAM to always have mini-batch larger than 1 sentence: `BatchNorm1D` works, although it normalizes only at mini/batch level, instead of batch level;
 - 1 GPU without enough RAM to always have mini-batch larger than 1 sentence: we are currently unable to cover this case.

In addition, `SyncBatchNorm` is not supported on CPU. So `BatchNorm1D` enables using the model on CPU.

# What changes does the patch introduce?
A flag parameter is added to control whether to use `SyncBatchNorm` or `BatchNorm1D`.

# How was this patch tested?
UT