[!41][SUBTITLING] Add script to make srt file starting from subtitled text, timestamp and yaml
# Why is the change needed?
The .srt file is the standard file used for subtitles that includes information about the text that has to be visualized and at which time. But there are no scripts for making the .srt file in our repository.

# What changes does the patch introduce?
Adds the .srt file maker which starts from the output produced by our models, using both the text and the timestamp (introduced by MR !14).

# How was this patch tested?
The obtained .srt has been tested on subtitling tools (VLC and online tools).