[!43][BUG][TESTS][CTC] Test passes even when doctests fail
The test for CTC based on doctest currenlty passes even though doctest tests are failing, because it just prints the failure but throws no assertion error.

Checks for failures, so the test fails when there are doctests not passing.

before fix, making doctest failing lets test pass. After the fix it does not.