[!44][NE] Determine tags at word level
# Why is the change needed?

When we predict the tags at token level, we risk that the same word is partly tagged with one NE type and another part with a different type. This is an impossible condition, so we should determine the most likely NE at word level.

# What changes does the patch introduce?

This patch determines the most likely NE type for each word while generating the output, hence avoiding the condition in which a wrod is partly tagged with a NE type and partly with another.

# How was this patch tested?

manual runs + added UTs