[!48][MISC] Log which samples cause OOM
# Why is the change needed?

Bad samples can cause OOM issues that are hard to understand. Namely, finding the sample which causes the OOM is like looking for a needle in a haystack. Identifying which samples are included in the batch causing the OOM can ease the life of the poor developer facing the OOM issue, significantly reducing the size of the haystack where to look for the needle.

# What changes does the patch introduce?

The patch adds a log showing the IDs of the samples in the batch that causes the OOM.

# How was this patch tested?
manual execution