[!35][CTC] Add best path decoding to the CTC generator (as sampling method)
# Why is the change needed?

To show some the effectiveness of CTC beam search, a comparison with the results obtained using greedy search can be useful.  For this reason, we add the possibility to generate the CTC output with greedy search as well.

# What changes does the patch introduce?

Add the greedy search generation, enabled with the `--sampling` command-line option.

# How was this patch tested?

manual tests