[!56][TRIANGLE] Correct config lm path parameters error arising when a newer environment is used
# Why is the change needed?
With the new environment (needed for A40 GPUs) and successive python libraries versions, the primary and auxiliary LM paths are inserted in the task configuration, raising an error when the generate script search for this information in the generic config

# What changes does the patch introduce?
If the paths are searched in the generic config and are not found, then they are searched in the task config.

# How was this patch tested?
Manual running in the new environment.