[!57][SUBTITLING] Corrections to CTC align script to make timestamp more subtitle-like
# Why is the change needed?
This PR: 1) address the error rising when the ground truth aligment matrix is longer than the audio file, and 2) correct the timing returned by the built-in ctc segmentation library since it is not suited to compute subtitles (it computes timestamp by averaging the feature duration instead we need a more conservative solution to avoid cutting too early or too late the subtitles).

# What changes does the patch introduce?
Correct the error 1) by returning start and end time of the segment while raising a warning and add a script based on the CTC alignment that produces ad-hoc timing for subtitling.

# How was this patch tested?
Manual runs.