[!63][CONFORMER] Correction to the Convolutional Layer of Conformer for missing padding
# Why is the change needed?
The Conformer model has a padding problem in the Convolutional Layer that is more evident at inference time when large batches are used: the output differs from the one obtained with batch size 1.
This is given by the fact that convolution and batch norm layers do not preserve padding.

# What changes does the patch introduce?
Apply padding in the Convolutional Layer after each aforementioned input transformation.

# How was this patch tested?
UT and check of the output by manual runs.