[!66][SUBTITLING] Subtitle timestamps through Levenshtein distance
# Why is the change needed?
At the moment, there is no way to obtain timestamp information for the target translation.

# What changes does the patch introduce?
This PR introduces a script that, given the textual subtitles in the source and target languages and the timestamp in the source language subtitle, computes the timestamp for the the target translation in subtitle form by exploiting Levenshtein distance and optimal alignment.

# How was this patch tested?
UT and manual runs.