[!67][NE] Add tags prediction to S2T Conformer models
# Why is the change needed?

In the current codebase, the prediction/generation of tags together with the target text is supported only for dual-decoder models.
This patch adds the feature to single-decoder models and specifically it includes a Conformer model that predicts tags in addition to the text.

# What changes does the patch introduce?

Adds a new model to predict tags with a S2T conformer model and implements the task and generation to support models that predict both text and tags.

# How was this patch tested?

UTs and manual experiments