[!70][CONFORMER][BUG] Fix padding on Conformer attention
# Why is the change needed?

In the Conformer Attention module of the Conformer Encoder, the output of the attention score was not padded. Even if it does not produce significant differences in the final predictions, the correction should be applied.

# What changes does the patch introduce?

The encoder padding mask is passed during the attention computation and it is applied after the attention score calculation.

# How was this patch tested?

UT and manual runs.