[!72][API] Simple IO server
# Why is the change needed?

We need to expose the processing capabilities of our models to the external worlds.
As a first simple showcase, we introduce a simple server, which accepts communication over IO.
Such a server accepts JSON lines in stdin and returns the output of the model in stdout, always in JSON format.

# What changes does the patch introduce?

Adds the simple IO server.

# How was this patch tested?

Manual runs and UTs.