[!78][DEV] Add release template
# Why is the change needed?

When we release the code open source we need to add a README and instructions for the maintainer of the public repo on what needs to be released.
This MR template is not a good fit for this specific MRs, as in these cases the motivation is obvious,
and the information needed (e.g. the list of commits related to the release) is different from normal MRs.

# What changes does the patch introduce?

Introduces a new MR template to use when creating MRs for releasing open source the code.

# How was this patch tested?

NA