[!75][DUALENC] Add support for multimodal encoder (speech and text)
The current `S2TTransformerModel` only supports one encoder (speech encoder) while having another encoder would be useful to exploit different modalities (e.g. text) or to encode different audio information.

Adds another encoder to the `S2TTransformerModel` which is composed of a stack of Transformer encoders. The cross-attention with the `TransformerDecoder` can be computed in a parallel way (two cross-attentions are computed individually between `TransformerDecoder` and the two encoders and then summed together) or in a sequential way (the first cross-attention is computed and then passed as input to the second cross-attention computation), as proposed in [Gaido et al. 2020](https://www.isca-speech.org/archive/pdfs/interspeech_2020/gaido20_interspeech.pdf).

Manual training and UTs.