[!81][MISC] Add script to produce multilingual SentencePiece vocabulary
# Why is the change needed?
Sometimes it is useful to generate a SentencePiece model which is trained on many languages but, to date, the repository does not provide a script to generate it.

# What changes does the patch introduce?
This MR introduces a script that automatically generates a SentencePiece model starting from tsv files we already have from the ST preprocessing.

# How was this patch tested?
Manual run.