[!83][MISC] Re-organize README papers
# Why is the change needed?

The list of papers is growing and is becoming hard to find a specific paper. In addition, the current order puts new papers at the bottom, while they are more likely to be interesting, so they should be on the top.

# What changes does the patch introduce?

The MR divides the papers by year and adds the venue to them.

# How was this patch tested?

NA