[!73][CONFORMER][BUG] Implement relative shift robust to padding
# Why is the change needed?

The current implementation of the Conformer architecture is not robust to padding because the encoder returns different results for the same input according to whether it is padded or not. This means that inference with different GPUs or batch sizes leads to different results, as well as training can be different according to the batch size. The only remaining problem in the current implementation was the relative shift operation.

# What changes does the patch introduce?

The patch introduce a padding-aware and safe implementation of the relative_shift operation. It also adds a flag to restore the old implementation as the introduced fix causes a significant slowdown of the training (4 min every 100 updates instead of 2:30 on 4 A40).

# How was this patch tested?

added UTs, manual runs