[!87][SIMULTANEOUS] Add Local Agreement agent
# Why is the change needed?
The Local Agreement is the last winning policy of the IWSLT Simultaneous Speech Translation evaluation campaign (2022) that is applied to offline-trained systems to obtain simultaneous inference. Comparisons with this policy are needed for benchmarking, however, this agent is not currently present in our repository.

# What changes does the patch introduce?
Add the Local Agreement agent based on the IWSLT 2022 docker adapted to FairseqST.

# How was this patch tested?
Manual runs.