[!88][DEV] Refactor to isolate libraries specific for speech translation
As people might need to work on projects not related to speech translation,
we want to avoid adding dependencies that are needed only for speech translation project
in the main setup file.
Conversely, we do not want to make difficult the installation for speech projects.

The patch creates a dedicated requirements txt file that contains the dependencies specific for ST projects and
moves the speech-only libraries to the new requirement file.

Existing CI