[!91][SIMULTANEOUS] EDAtt agent for Simultaneous Speech Translation
# Why is the change needed?
Up to now, only the wait-k and Local Agreement policies are available for SimulST. However, in our recent paper (Attention as a Guide for Simultaneous Speech Translation: https://arxiv.org/abs/2212.07850), we prove that an adaptive policy named EDAtt which exploits the attention scores to guide the simultaneous inference is not only better in terms of quality but also in terms of computational efficiency.

# What changes does the patch introduce?
It adds EDAtt agent (compatible with SimulEval) to the simultaneous speech translation agents.

# How was this patch tested?
Manual runs and UT.