[!94][SIMULTANEOUS][BUG] Correct agents after refactoring
# Why is the change needed?
Wait-k and EDAtt agents do not work after refactoring in MR !91 due to some errors in calling 1) the `super()` methods, 2) the `_predict` method was called but not exists anymore, and 3) EDAtt misses the `predict` method (before refactoring, it was inheriting this method from the LocalAgreement). Moreover, the `new_segment` and `write` parameters in `states` were not initialized in the Base agent.

# What changes does the patch introduce?
Corrected the error in the `super()` call, and refactored the `generate_hypothesis` in the Base Agent by including the `_get_prefix` and `_get_prefix_len` which are now in common among all the agents, Wait-k included. Also unified the `policy` and `predict` methods among all the agents to avoid useless duplicates. In doing this, the policy of the Wait-k has been completely refactored. Lastly, corrected the initialization error in the `initialize_states` and unified it in the Base agent to avoid duplicates. Also, introduced a base test class for the SimulST agents, inherited by the specific policies and which tests common functionalities among the various implemented policy.

# How was this patch tested?
Manual runs and UT for the Wait-k policy.