[!99][GENDER] Add MuST-SHE coverage and gender accuracy evaluation script
# Why is the change needed?

The script to compute gender accuracy and term coverage is currently not included in our repository.
Although it is present in the MuST-SHE release and in other public repositories, adding it here helps its re-usage
and we may point to this new version in future works.

# What changes does the patch introduce?

Add the MuST-SHE evaluation scripts, which has been slightly re-written to improve its readability and reusability.

# How was this patch tested?
added UTs