[!100][SIMULTANEOUS] Correct Local Agreement error with language prefix
# Why is the change needed?
Local Agreement agent for Simultaneous ST cut the first word of the prediction two times (in the BaseAgent and in the `_policy` of the LocalAgreementAgent) during inference when `prefix_token_idx` is enabled (multilingual case).

# What changes does the patch introduce?
Correct the error and unify the `prefix_token_idx` handling with the Base SimulST agent as for the other implemented agents.

# How was this patch tested?
UTs and manual runs.