[!98][MISC] Add task for extra tokens
# Why is the change needed?
Currently, it is impossible to fine-tune a model using an extended target vocabulary, while it can be particularly useful if we want to exploit a pre-trained model and adapt it to a new sequence-to-sequence task (e.g., train a model for Simultaneous ST to produce translated tokens or a tag `<READ>` meaning that we have to wait for more input audio).

# What changes does the patch introduce?
Add a task in which a series of tokens can be added to the target vocabulary and initialize their weights with the (standard) normal distribution.

# How was this patch tested?
UT and Manual runs.