[!51][MULTITASK] Add implementation of multi-task with auxiliary classifier
# Why is the change needed?

In planned works we aim at classifying a feature on the encoder output either to retain it or to avoid that it is extracted by the network. In the latter case, we use a gradient reversal over the classifier.

# What changes does the patch introduce?

The patch add a first implementation of a multitask system that performs both the speech translation and a classification from speech. Specifically, we introduce:

 - a dataset reading the `aux_class` field in the TSV dataset;
 - a task using the above-mentioned dataset;
 - two models based respectively con Conformer and Transformer to execute both tasks;
 - a gradient reversal layer that can be enabled to "hinder" the classified feature extraction;
 - a script to add the preferred pronoun as `aux_class` to the TSV dataset of MuST-C.

# How was this patch tested?

Added UT + manual tests