[!110][BUG] Avoid BatchNorm is updated when freezed
# Why is the change needed?

When freezing a part of an architecture that contains a `BatchNorm`,
we are currently still updating the variables in `BatchNorm` as they have to be set
to evaluation mode, but before each epoch fairseq sets a model in training mode.

In this way, models which should be freezed are updated with respect to the batch normalization layers.

# What changes does the patch introduce?

Overrides the `train()` function to avoid that `BatchNorm` are set in training mode
when they should be freezed.

# How was this patch tested?

updated existing UTs to reflect fairseq behavior