[!120][AUDIO] Add support to store/read waveforms and extract filterbanks on the fly
# Why is the change needed?

In some use cases, we might want to apply transformations to the waveform before they are converted into spectrograms. This is requires if e.g. we want to alterate the waveform in different iterations of the training.

# What changes does the patch introduce?

The patch adds the support to write waveforms instead of spectrograms in the `preprocess_generic` script, to read them and apply transformations on them for the `speech_to_text` tasks.

# How was this patch tested?

UTs + manual experiment: a model trained on the spectrograms obtains the same results on the same test set with the new preprocessing and reading