[!125][BUG][SUBTITLE] SRT to blocks add EOL to empty lines
# Why is the change needed?

To compute metrics such as Sigma, we need to convert SRT files into lines with `<eol>` and `<eob>`.
The script doing so, as of now, adds `<eol>` even when there is an empty line in the SRT, which means that the output of the script contains lines ending with `<eol> <eob>` that does not make sense.

# What changes does the patch introduce?

Fixes the behavior, so trailing empty lines are not considered to add `<eol>`.

# How was this patch tested?

added test case