[!130][SUBTITLING][BUG] When text is longer than audio, alignment script should support more than one eob
# Why is the change needed?

If the text to be aligned is shorter than the source (in terms of number of BPE vs audio frames),
the CTC alignment cannot be performed, and we output a single segment of size equal to the audio length.
However, if there is more than one block in the generated text, this behaviour creates an inconsistency,
as the number of segments in the text and in the generated times is different.

# What changes does the patch introduce?

Splits the audio in segments of equal length when the alignment cannot be performed, instead of assuming a single segment.

# How was this patch tested?

manual runs