[!131][SUBTITLING][BUG] Ensure last block is not ignored in alignments
# Why is the change needed?

The current script for Levenshtein alignment may end up considering
as an insertion the last block in the subtitle. This happens when
the subtitle contains one block more than the transcript at the end
and the subtitle is a bit longer than the transcript.

# What changes does the patch introduce?

The patch ensures that the alignments always put the last end of block
at the end, and in case the last block is considered an insertion after
the last block, the last block is moved at the end.

# How was this patch tested?

UTs + manual runs