[!132][SUBTITLING] Assert nothing is discarded when creating SRT files
# Why is the change needed?

The script that creates SRTs currently assumes that the number of timestamp segments and textual blocks are the same.
This assumption, however, is never checked, and in case it does not hold, only the common elements from the timestamp segments and textual blocks is processed, while the rest is ignored.
This means that, if the number of timestamp segments is lower than the number of text blocks, all text blocks
for which the timestamps are missing are ignored and the generated SRT will not contain them.

# What changes does the patch introduce?

Adds an assertion that exits with error if the above assumption does not hold.

# How was this patch tested?

manual runs