[!129][MISC] Add support to extract attention scores for speech-to-text models
# Why is the change needed?

Attentions weights can be used to know the alignment between the generated tokens and the source audio.
This can have several applications, ranging from simultaneous to subtitling.

# What changes does the patch introduce?

Adds the support to attention weights printing for the speech-to-text task, and prints
CTC compressions, so that the alignment with the source audio can be inferred.

# How was this patch tested?

manual runs