[!141][MISC] Add pretrained model loading and freezing for multi-task models
# Why is the change needed?

Multi-task models do not support pretrained model loading and freezing,
unless for the case of a multi-task model with classifier.

However there are other scenarios in which this feature is important
for multi-task models (e.g., for multi-task models with joint CTC on target).

# What changes does the patch introduce?

Enables pretrained model loading and freezing for all multi-task models.

# How was this patch tested?

manual runs