[!123][SUBTITLING] Add training with CTC loss on target text
# Why is the change needed?

With the goal of having a real end-to-end subtitling model, we should avoid relying on intermediate transcripts. We can instead predict timings directly on the target side with a CTC loss predicting the target text.

# What changes does the patch introduce?

Adds a model and criterion to compute CTC loss on the target text.

# How was this patch tested?

added UTs