[!153][SUBTITLING] Avoid repeated <eob>/<eol> during generate
# Why is the change needed?

If subtitle blocks and lines cannot be empty. As such, if a model predicts consecutive `<eob>` and/or `<eol>`
the resulting output is not a valid subtitle, leading to invalid SRT that cannot be used/evaluated.

# What changes does the patch introduce?

Introduces a new command line option to enable forcing that `<eob>` and/or `<eol>` cannot be
generated one after the other without intermediate content.

# How was this patch tested?

Added UTs