[!155][STREAMING] StreamAtt decision policy
# Why is the change needed?
Currently, it is not possible to emit translation in real-time by receiving a continuous stream of audio. Indeed, SimulST agents keep all the history (ie. past audio and text), so, if the audio is not pre-segmented in advance, they become extremely slow and fail for out-of-memory.

For this reason, we need streaming agents that determine not only when and what to emit, but also what to keep from the past and what can be discarded.

# What changes does the patch introduce?

It introduces a new `StreamingSTAgent` that performs streaming speech translation in the following steps:
 - it leverages a `BaseSimulSTAgent`, ie. a simultaneous ST agent, to determine the `policy` that decides when and how many tokens to emit;
 -  it leverages a `HistorySelection` class to determine which text and audio from the past has to be retained and which has to be discarded.

As `HistorySelection` mechanism, we introduce two methods:
 - `FixedWordsHistorySelection` that keeps a fixed amount of past words and their corresponding audio (using the attention to align the text and audio);
 - `PunctuationHistorySelection` that keeps all the words after the last strong punctuation mark, and the corresponding audio (similarly determined through the attention).

# How was this patch tested?
UTs and manual runs.