[!171][XAI][BUG][REFACTOR] Move perturbators and scorers registering into dedicated folders
# Why is the change needed?

The current flat structure of the `occlusion_explanation` folder does not allow for intuitively adding new scorers/perturbators without import errors.

With this refactoring, adding a new scorer/perturbator is as easy as add a new file with the proper register function, as it is for criterions, models and so on.

# What changes does the patch introduce?

Creates two dedicated packages where scorers and perturbators are moved.

# How was this patch tested?

NA