[!170][XAI] Improve initialization from config
# Why is the change needed?
The code within the `get_perturbator_from_config()` function in the configuration is redundant. Additionally, in the event of a misspelling in the YAML file, default values are used, yet the user may not be aware of this occurrence.

# What changes does the patch introduce?
To eliminate redundancy, the assertion is removed, and a logger info message is included to notify the user of the availability of the occlusion choice in the YAML file.

# How was this patch tested?
Manual run.