[!173][XAI][BUG] Fix n_masks type
# Why is the change needed?
In the occlusion dataset class, the number of masks is derived from the encoder perturbator class. However, if the perturbator is based on SLIC segmentation, the number of masks can be adjusted to be equally divided by each segmentation layer. Consequently, this operation results in the `n_masks` variable being assigned a float value. This leads to errors in certain dataset methods that use `n_masks` in `range()`.

# What changes does the patch introduce?
The `n_masks` variable is modified to ensure it is of integer type.

# How was this patch tested?
Manual run.