[!172][XAI] Dynamic slic segmentations
# Why is the change needed?
In SLIC segmentation, the number of segments is fixed, regardless the audio duration. Moreover, the default values are not aligned to [previous XAI work](https://arxiv.org/pdf/2006.02659.pdf) using SLIC nor to the [SLIC documentation](https://scikit-image.org/docs/stable/api/skimage.segmentation.html).

# What changes does the patch introduce?
Parameters related to SLIC segmentations have been updated. It is now feasible to specify the number of segments for each segmentation layer relative to a reference duration, which is subsequently used to adjust the segment count based on the duration of each individual sample.

# How was this patch tested?
UTs.