[!177][XAI][BUG] Fix resolution of max_positions
# Why is the change needed?

There is a bug in the resolution of max positions in XAI generate scripts.
As in the case of SpeechToText the max positions of the model (and task) are a tuple (source, target) and the current code expands the tuple of the model max_positions, the utils is ignoring the max_position of the source and always uses the max positions of the target.

# What changes does the patch introduce?

Avoids that the tuple returned by the model max_positions is expanded when computing the overall max_positions.

# How was this patch tested?

manual runs