[!180][XAI] Allow 0.0 as perturbation probability in occlusion-based methods
# Why is the change needed?

Currently the initialization of the occlusion based methods from the config file with p = 0 (i.e., no perturbation) is not allowed.
As such we cannot use occlusions methods that perturb either the input only or the previous output tokens only.

# What changes does the patch introduce?

Give the possibility to set a p value equal to zero (i.e., no perturbation). This might be useful in cases where we want to perturb only filterbank and not tokens, or viceversa.

# How was this patch tested?
Manual run