[!187][XAI] Add Aggregators
# Why is the change needed?
Currently, only sentence-level aggregation of explanations is supported. To perform some analyses, other types of aggregations could be useful, such as aggregations of explanations related to multiple tokens (e.g., to obtain word-level explanations) or aggregations of explanations over the time or frequency dimensions.

# What changes does the patch introduce?
Adding three new aggregators:
- Phrase-level aggregator, which aggregates explanations of multiple tokens into higher level linguistic units, such as words.
- Time-level aggregator, which aggregates scores for both filterbank and previous output tokens explanations over the time dimensions.
- Frequency/Embedding-level aggregator, which aggregates explanations by collapsing the channel/embedding dimension (in case of discrete explanations, this aggregation is redundant).

Moreover, some adaptations to the available sentence-level aggregator (and consequently to the metrics which use such aggregator) have been made, in order to make it compatible with the new aggregators, and to also allow to use multiple aggregators together.

# How was this patch tested?
UTs added.