[!188][XAI] Slic N. Segments
# Why is the change needed?
Currently, the configuration file for the perturbator that uses the SLIC segmentation in occlusion-based explanations requires a tuple for the number of segments attribute. However, this can be convoluted, requiring a syntax similar to `n_segments: !!python/tuple [1000]` in the configuration file.

# What changes does the patch introduce?
Change the type of `n_segments` attribute.

# How was this patch tested?
Modifying available UTs.