[!199][XAI] Load original probs on the fly
# Why is the change needed?

In the current implementation, the original probabilities for all the sample are loaded from an h5 file before starting the iteration over batches. To reduce memory consumption, this operation can be performed during the iteration over batches loading only the samples involved in the current batch.

# What changes does the patch introduce?

Adjusting the original probabilities loading.

# How was this patch tested?

Manual running.