[!210][XAI] Add Possibility of Saving Scores in Ablation Metric
# Why is the change needed?
The current implementation of ablation metric compute the WER/BLEU scores for each step of ablation. If a specific argument is provided, it also plots the graph of the scores obtained. However, it does not save the scores, which could be useful when one wants to plot more scores together in the same graph. Therefore, this MR aims at introducing the possibility of also saving the scores.

# What changes does the patch introduce?
In the cli script, an argument is added and scores can be saved as numpy files.

# How was this patch tested?
Manual running.