[!213][XAI] Policy for arbitrating between a fixed or dynamic number of SLIC clusters
# Why is the change needed?
Setting the number of clusters dynamically, in a way proportional to the utterance length, is only beneficial up to a certain point. If the utterance is too long, the number of clusters can become too large with respect to the number of masks, so each cluster is not masked enough times to generate good-quality explanations. This patch introduces a policy to arbitrate between the two approaches, setting a cap in utterance length over which the dynamic strategy is replaced with a fixed one.

# What changes does the patch introduce?
Adds the parameter `threshold_duration` to the SLIC pertubator config. If an utterance's number of frames is above this threshold, the number of clusters remains fixed instead of being set dynamically.

# How was this patch tested?
Added unit tests.