[!200][SIMULTANEOUS] SimulSeamless: simultaneous ST agent for SeamlessM4T
# Why is the change needed?
SeamlessM4T is a family of pre-trained models with multimodal input and output (text and speech). In particular, the speech-to-text model covers more than 143 source languages and 200 target languages, realizing one of the most multilingual models available. However, to provide outputs in a simultaneous manner without modifying the architecture, a SimulST agent should be applied.

# What changes does the patch introduce?
SimulST agents (wait-k and AlignAtt) for the speech-to-text SeamlessM4T model are introduced. The models are implemented on HuggingFace transformers.

# How was this patch tested?
UTs and manual runs.