[!222][ARCH] Decoder-only speech-to-text model
# Why is the change needed?

The current trend of integrating speech encoders with pretrained large language models rely on architectures that are different from the traditional encoder-decoder attention-based models. Specifically, instead of the cross-attention between the decoder embedding and the encoder output, these solutions rely on concatenating the encoder output with the decoder embeddings, eliminating the cross-attention mechanism. In literature, this architecture has often been referred to as decoder-only.

# What changes does the patch introduce?

Adds both Conformer-encoder and Transformer-encoder architectures that rely on encoder output prepending instead of cross-attention.
For these architectures we include two variants: with or without causal masking on the speech encoder output.

# How was this patch tested?

added UTs + manual runs