[!229][BUG][ARCH] Fix decoder-only/decoder-prepending implementation
# Why is the change needed?

The implementation of decoder-only and decoder-prepending introduced in !222 (63d205f709503979d47552a367a1dc281f36b268) does not work properly as it does not support incremental decoding and does not account for the fact that the self-attention mask is cached. The latter point leads to the issue that tokens will be able to look into the future if previously a longer audio has been processed, compromising the training.

# What changes does the patch introduce?

Fixes the self-attention mask handling, also introducing the option to use causal masking in the audio prepended to the decoder. Also, it fixes the incremental decoding process, used in the generate.

# How was this patch tested?

Added UTs