[!233][XAI][BUG] Division by zero in the mean standard normalizer
# Why is the change needed?
If the saliency scores in an explanation are very similar to each other, problems can arise when performing mean normalization. In some very rare cases, if all values are the same, the standard deviation will be zero which leads to an error (dividing by zero).

# What changes does the patch introduce?
When a division by zero occurs, the normalized saliency scores for those features are set to 0.

# How was this patch tested?
Manual tests on an occurrence of the very rare case described above.